requirements_core.txt
# Core Dependencies (Required)
aiohttp>=3.9.0
requests>=2.31.0
pyyaml>=6.0.1
psutil>=5.9.6

# Monitoring & Database
prometheus-client>=0.19.0

# Security & Encryption
cryptography>=41.0.7

# Performance
# uvloop>=0.19.0  # For faster asyncio (Linux/Mac only)
aiofiles>=23.2.1  # For async file operations

# Utilities
python-dotenv>=1.0.0
colorlog>=6.8.0
tqdm>=4.66.1
